refactor: remove unused addon modules and update SQLAlchemy configuration
