--- hypern/database/addons/__init__.py
@@ -1,5 +1,0 @@
-from . import sqlalchemy
-
-__all__ = [
-    "sqlalchemy",
-]
