--- hypern/database/nosql/addons/unicode.py
@@ -1,10 +1,0 @@
-from mongoengine import StringField
-
-
-class UnicodeField(StringField):
-    def validate(self, value):
-        try:
-            value.encode("utf-8")
-        except UnicodeEncodeError:
-            self.error("Value must be valid Unicode")
-        return True
