+++ pyproject.toml
@@ -22,16 +22,13 @@ dependencies=[
     "sentry-sdk==2.11.0",
     "pydantic-settings==2.3.4",
     "celery==5.4.0",
-    "asgiref==3.8.1",
     "psycopg==3.2.3",
     "pyyaml == 6.0.2",
-    "mongoengine==0.29.1",
     "argon2-cffi==23.1.0",
     "bcrypt==4.2.0",
     "orjson==3.10.11",
     "multiprocess==0.70.17",
     "uvloop==0.21.0; sys_platform != 'win32' and platform_python_implementation == 'CPython' and platform_machine != 'armv7l'",
-    "cryptography==43.0.3",
     "watchdog==6.0.0",
     "jsonschema==4.23.0",
     "psutil==6.1.0",
@@ -58,17 +55,14 @@ pydash = "^8.0.3"
 sentry-sdk = "^2.11.0"
 pydantic-settings = "^2.3.4"
 celery = "^5.4.0"
-asgiref = "^3.8.1"
 psycopg = "^3.2.3"
 maturin = "^1.7.4"
 pyyaml = "^6.0.2"
-mongoengine = "^0.29.1"
 argon2-cffi = "^23.1.0"
 bcrypt = "^4.2.0"
 orjson = "^3.10.11"
 multiprocess = "^0.70.17"
 uvloop = { version = "^0.21.0", markers = "sys_platform != 'win32' and (sys_platform != 'cygwin' and platform_python_implementation != 'PyPy')" }
-cryptography = "^43.0.3"
 watchdog = "^6.0.0"
 aiohttp = "^3.11.10"
 jsonschema = "^4.23.0"
